Add update attendance action.
