+++ client/app/bundles/Events/actions/index.jsx
@@ -1,34 +1,2 @@
-import axios from 'axios';
-
-import {
-  FETCH_EVENTS,
-  FETCH_EVENT,
-  FETCH_ATTENDANCES
-} from './types';
-
-export const fetchEvents = (queryString = '') => {
-  const request = axios.get(`/events.json${queryString}`);
-
-  return {
-    type: FETCH_EVENTS,
-    payload: request
-  };
-};
-
-export const fetchEvent = (eventId) => {
-  const request = axios.get(`/events/${eventId}.json`);
-
-  return {
-    type: FETCH_EVENT,
-    payload: request
-  };
-};
-
-export const fetchAttendances = (eventId) => {
-  const request = axios.get(`/events/${eventId}/attendances.json`);
-
-  return {
-    type: FETCH_ATTENDANCES,
-    payload: request
-  };
-}
+export * from './EventsActions';
+export * from './AttendancesActions';
