+++ client/app/bundles/Events/actions/types.jsx
@@ -1,3 +1,4 @@
 export const FETCH_EVENTS = 'FETCH_EVENTS';
 export const FETCH_EVENT  = 'FETCH_EVENT';
 export const FETCH_ATTENDANCES = 'FETCH_ATTENDANCES';
+export const UPDATE_ATTENDANCE = 'UPDATE_ATTENDANCE';
