+++ client/app/bundles/Events/reducers/AttendancesReducer.jsx
@@ -1,4 +1,6 @@
-import { FETCH_ATTENDANCES } from '../actions/types';
+import _ from 'lodash';
+
+import { FETCH_ATTENDANCES, UPDATE_ATTENDANCE } from '../actions/types';
 
 const INITIAL_STATE = [];
 
@@ -6,6 +8,14 @@ export default (state = INITIAL_STATE, action) => {
   switch (action.type) {
   case FETCH_ATTENDANCES:
     return action.payload.data.data;
+  case UPDATE_ATTENDANCE: {
+    const updatedAttendance = action.payload.data.data;
+    return _.map(state, function(attendance) {
+      return (attendance.id === updatedAttendance.id)
+        ? updatedAttendance
+        : attendance;
+    });
+  }
   default:
     return state;
   }
